2020/9/4

+++ PAT_Basic/1049 数列的片段和 (20分).cpp
@@ -10,9 +10,9 @@ int main(){
 	}
 	for(int i = 0; i < n; i++){
 		int index = i+1;
-		int sum1 = (n+(n-i))*index*0.5;
-		int sum2 = i*0.5*(1+i);
-		int sum = sum1 -sum2;
+		long long  sum1 = (n+(n-i))*index*0.5;
+		long long sum2 = i*0.5*(1+i);
+		long long sum = sum1 -sum2;
 		save+=a[i]*1.0*sum;
 	}
 	printf("%.2f",save);
